+++ verb.el
@@ -246,7 +246,10 @@ hierarchy."
 (defcustom verb-enable-elisp-completion nil
   "When set to a non-nil value, enable Lisp completion in code tags.
 Completion is handled by the `verb-elisp-completion-at-point'
-function."
+function.
+
+Note the the point must be between the two code tag delimeters
+\(e.g.  \"{{\" and \"}}\") for the completion function to work."
   :type 'boolean)
 
 (defface verb-http-keyword '((t :inherit font-lock-constant-face
@@ -432,6 +435,8 @@ more details on how to use it."
   :group 'verb
   (when verb-mode
     (verb--setup-font-lock-keywords)
+    (add-hook 'completion-at-point-functions #'verb-elisp-completion-at-point
+              nil 'local)
     (when (buffer-file-name)
       (verb--log nil 'I
                  "Verb mode enabled in buffer: %s"
@@ -606,14 +611,21 @@ KEY and VALUE must be strings.  KEY must not be the empty string."
 (defun verb-elisp-completion-at-point ()
   "Completion at point function for Lisp code tags."
   (when-let (verb-enable-elisp-completion
+             ;; Get the contents inside the code tag: {{<content>}}
              (beg (save-excursion
-                    (when (search-backward (car verb-code-tag-delimiters) (line-beginning-position) t)
+                    (when (search-backward (car verb-code-tag-delimiters)
+                                           (line-beginning-position) t)
                       (match-end 0))))
              (end (save-excursion
-                    (when (search-forward (cdr verb-code-tag-delimiters) (line-end-position) t)
+                    (when (search-forward (cdr verb-code-tag-delimiters)
+                                          (line-end-position) t)
                       (match-beginning 0)))))
+    ;; Set up the buffer where we'll run `elisp-completion-at-point'
     (unless verb--elisp-completion-buffer
-      (setq verb--elisp-completion-buffer (get-buffer-create " *verb-elisp-completion*")))
+      (setq verb--elisp-completion-buffer
+            (get-buffer-create " *verb-elisp-completion*")))
+    ;; Copy the contents of the code tag to the empty buffer, run
+    ;; completion there
     (let* ((code (buffer-substring-no-properties beg end))
            (point-offset (1+ (- (point) beg)))
            (completions (with-current-buffer verb--elisp-completion-buffer
@@ -621,6 +633,8 @@ KEY and VALUE must be strings.  KEY must not be the empty string."
                           (insert code)
                           (goto-char point-offset)
                           (elisp-completion-at-point))))
+      ;; The beginning/end positions will belong to the other buffer, add
+      ;; `beg' so that they make sense on the original one
       (when completions
         (append (list (+ (nth 0 completions) beg -1)
                       (+ (nth 1 completions) beg -1))